pyproject.toml
[project]
name = "aws-mcp-proxy"
version = "1.0.0"
description = "AWS MCP Proxy for Microsoft Copilot Studio"
authors = [{ name = "Your Organization", email = "contact@example.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "httpx>=0.25.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
include = ["main.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
